python/requirements.txt
pan-os-python>=1.11.0
pan-python>=0.17.0
boto3>=1.34.0
botocore>=1.34.0
requests>=2.31.0
pyyaml>=6.0.1
jinja2>=3.1.4
tabulate>=0.9.0
rich>=13.7.0
pytest>=8.2.0
pytest-mock>=3.14.0
